Final Cleanup
That should be the final cleanup of the settings menu script, everything fires once when opened still but all settings load properly and it is limited as much as possible, encapsulated by it's own limitations, hopefully unable to break it's boundaries and cause me further headache.

I will continue trying to reduce variables and blueprint nodes where possible.